Implement nvim-lsp-installer install/uninstall, added more laps, defered startup, improve user feedback
