+++ contribute/Dockerfile
@@ -24,6 +24,9 @@ RUN pacman -Sy neovim --noconfirm
 # Install runtime dependencies of doom-nvim
 RUN pacman -Sy ripgrep nodejs-lts-fermium npm git bash gcc jq --noconfirm
 
+# Required for nvim-lsp-installer
+RUN pacman -Sy wget unzip
+
 # Required for watching Doom-nvim-contrib for changes
 RUN npm i -g chokidar-cli
